Move a bunch more things into util package

+++ src/main/scala/coreplex/BaseCoreplex.scala
@@ -11,7 +11,7 @@ import uncore.devices._
 import uncore.util._
 import uncore.converters._
 import rocket._
-import rocket.Util._
+import util._
 
 /** Number of memory channels */
 case object NMemoryChannels extends Field[Int]
@@ -148,7 +148,7 @@ abstract class BaseCoreplexModule[+L <: BaseCoreplex, +B <: BaseCoreplexBundle](
       tile.interrupts.meip := plic.io.harts(plic.cfg.context(i, 'M'))
       tile.interrupts.seip.foreach(_ := plic.io.harts(plic.cfg.context(i, 'S')))
       tile.interrupts.debug := debugModule.io.debugInterrupts(i)
-      tile.hartid := i
+      tile.hartid := UInt(i)
       tile.resetVector := io.resetVector
     }
